requirements.txt
asn1crypto==0.24.0
attrs==19.1.0
autobahn==19.9.2
Automat==0.7.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
constantly==15.1.0
cryptography==2.7
dateparser==0.7.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.19
PyHamcrest==1.9.0
pyOpenSSL==19.0.0
python-binance==0.7.0
python-dateutil==2.8.0
pytz==2019.2
regex==2019.8.19
requests==2.22.0
service-identity==18.1.0
six==1.12.0
Twisted==19.7.0
txaio==18.8.1
tzlocal==2.0.0
urllib3==1.25.3
zope.interface==4.6.0
